refactor(jni): queryable refactor
